+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 6.4.8
 ### Fixes
 - `primary_nic` in `ionoscloud_server` should be set on creation
+- `ssh_keys` was no longer being set if server was not vcpu.
+- `ssh_keys` will no longer be computed on any type of server
+- `ssh_key_path` will now be set to schema on creation
 
 
 ## 6.4.7